+++ compilerAst/src/prog8/compilerinterface/IStringEncoding.kt
@@ -1,6 +1,6 @@
 package prog8.compilerinterface
 
 interface IStringEncoding {
-    fun encodeString(str: String, altEncoding: Boolean): List<Short>
-    fun decodeString(bytes: List<Short>, altEncoding: Boolean): String
+    fun encodeString(str: String, altEncoding: Boolean): List<UByte>
+    fun decodeString(bytes: List<UByte>, altEncoding: Boolean): String
 }
