+++ compilerInterfaces/src/prog8/compilerinterface/ICompilationTarget.kt
@@ -3,6 +3,6 @@ package prog8.compilerinterface
 interface ICompilationTarget: IStringEncoding, IMemSizer {
     val name: String
     val machine: IMachineDefinition
-    override fun encodeString(str: String, altEncoding: Boolean): List<Short>
-    override fun decodeString(bytes: List<Short>, altEncoding: Boolean): String
+    override fun encodeString(str: String, altEncoding: Boolean): List<UByte>
+    override fun decodeString(bytes: List<UByte>, altEncoding: Boolean): String
 }
